Add the field "rbac_actions" to the model query hooks
This patch adds a new field to the models query hook: "rbac_actions".
This field stores the multiple RBAC actions to filter the query in
case of making a non-admin query.

The default RBAC action are "access_as_shared" and "access_as_readonly".
For networks and subnets, the action "access_as_external" is needed in
this query. Instead of adding a query hook, this additional field allows
to modify the required RBAC action filter.

Related-Bug: #2059236
Change-Id: Ic41461223719c535ac9d2bbadfb1b6077c898903