Fix incorrect links for the filtering spec
The spec networking_general_api_information.html was moved
to api-ref in [0] but some of the links pointing to it
were not updated. This change makes all the refs consistent
and adds links where missing.

Also moved sync/async and bulk create sections lower to
group all the filtering, pagination and sorting sections
together since they are inter-related.

[0] https://review.opendev.org/c/openstack/neutron-lib/+/362877

Change-Id: I895cc9aaee2546b95e93150c134f350a50346447