Catch all exceptions in convert_ip_to_canonical_format
Recent netaddr started to raise AttributeError for subnetpools values
like [{'start': ..., 'end': ...}].

The converter is meant to be graceful towards inputs that are not proper
ip addresses.

Instead of adding another exception type to the list and hope that the
list of expected types covers everything netaddr may raise, this patch
makes the converter catch all exceptions.

Change-Id: I967032c1dea690cd9da95ad3a8e382671186d971