tests: return same message for validate_[subnet*|route_cidr]
In netaddr 1.0.0, more requirements are enforced for input cidrs. This
makes the library raise exceptions for some values that were accepted
before and that we then manually validated (e.g. for the fact that the
library normalized value is identical to input.)

This patch makes the validators behave identically for both netaddr
versions (returning the same messages.)

This fixes some test failures with netaddr 1.2.1.

Change-Id: I4bc9c4238ed26cdc7ebf6e0c877ea4bd9c79c36d