api: canonicalize ipv6 addresses in `subnet:allocation_pools` field
Before the patch, `start` and `end` addresses in `allocation_pools` were
never canonicalized because an invalid converter was used for this field
that assumed a string as an input format, and not a list of dicts with
`start` and `end` keys, which is the correct format for the field.

The original invalid converter was added in
Ibef0bd9a2f7b7dd8ddccc183f17aa31e426c7fd0.

Related-Bug: #1531103
Change-Id: If768ab6a5f92d200841a734087bbc8fba8870dc6