Always expect project_id and tenant_id to be in the attributes list
Even in case when resource definition don't have project_id nor
tenant_id in its API definition it is always there because it is added
from the neutron_context.
Because of that reason, it was necessary to add those attributes to the
list of attributes of all resources, even those which don't really
belong to any specific project (like e.g. flavors or QoS rules).

This patch adds project_id and tenant_id to the list of the expected
attributes always so it won't fail validation of the attributes during
e.g. creation of the flavor even if flavor resource don't have
project_id defined.

Additionally this patch removes project_id from the flavors and QoS
rules definitions.

Closes-bug: #2022043
Change-Id: Ib25295e666e7230f9d6c7c5df83411014d5b6ac2