[api-ref] Add note about no validation of the target_tenant in RBAC API
Neutron can't really validate if string provided as 'target_tenant' in
the /v2.0/rbac-policies POST/PUT request is project ID or something
else, like e.g. project's name.
This patch adds small note about this to the Neutron api-ref document so
that users can be more aware of that.

Related-Bug: #2065743
Change-Id: Ie5bec404c372209ca15e4c8454957040032751cb