Add the ability to define the ``BaseWorker.set_proctitle`` flag
This patch allows to define the `BaseWorker.set_proctitle`` after
the initialization, needed to define the process title or not when
the process is started. When a ``BaseWorker`` is spawned in a
thread instead of a single process, this worker should not modify
the process title that should have been defined by the main thread.

That is happening, for example, with the workers spawned in the
``AllServicesNeutronWorker`` class. This class is defining the
process title as "neutron-server: services worker", but the new
workers spawned (``PeriodicWorker``) are re-writting the process
name to "neutron-server: periodic worker"

Change-Id: I744343804cd6b6a4bc7353502d7a0295800b37bd
Related-Bug: #2069595