Add new API extension ``uplink-status-propagation-updatable``
Add the API extension ``uplink-status-propagation-updatable`` to make
the port field ``propagate_uplink_status`` updatable. Now is possible
to send a port PUT request to update the value defined during the
creation.

This change will have an effect in the ML2/SR-IOV agent that will need
to update the VF status accordingly.

Related-Bug: #2078661
Change-Id: I96d8b1cfe2f0cce2198d74f48abfc90d952384bd