Do not define the executor of the RPC server
The method ``oslo_messaging.get_rpc_server`` allows to define the
executor between "eventlet" and "threading".

By default the server class used is
``oslo_messaging.rpc.server.RPCServer`` [1]. If no executor is defined,
the base class will get it depending on the context [2][3]. If
``thread`` is not monkey patched, the executor to use will be
"threading".

[1]https://github.com/openstack/oslo.messaging/blob/412ab4de9244d99903b7ada7a60ef19156cd8451/oslo_messaging/rpc/server.py#L222
[2]https://github.com/openstack/oslo.messaging/blob/412ab4de9244d99903b7ada7a60ef19156cd8451/oslo_messaging/server.py#L334
[3]https://github.com/openstack/oslo.messaging/blob/412ab4de9244d99903b7ada7a60ef19156cd8451/oslo_messaging/_utils.py#L87-L92

Change-Id: I1541137ea59cc4bba5699f2e98b80ed3ddbe5337
Related-Bug: #2069831