Use py312 for all neutron-lib jobs
There is a job definition here for testing
neutron-lib with the master branch of neutron.
As neutron is no longer testing py311 in its master
gate for 2025.1, start using the py312 version
of this jobs so we can move forward. Changed it
to voting as it has been non-voting since
June 2024 and checking the past few months
changes have not shown any failures.

Also moved the sqlalchemy job to start using py312
as well.

Also updated the testing runtime to include 3.12
and removed 3.8 based on 2025.1 governance [0].

[0] https://review.opendev.org/c/openstack/governance/+/926150

Change-Id: Ife1abbf48c15a2440c8b1f5316c541781964ce5c