Fix ``external-gateway-multihoming`` API extension definition
The new field ``external-gateways`` added in the API extension
``external-gateway-multihoming`` didn't have the ``enforce_policy`` flag
defined and the validate parameters didn't provide the needed
information to the Neutron policy to build a correct rule match.
Now this field copies the ``ext-gw-mode`` extension validator used in
the ``external_gateway_info`` field.

The validator type ``list_of_dict_or_nodata`` is currently not
recognized by the Neutron policy as an iterable validator [1]; this
code must be changed in Neutron in order to accept this new defined
validator that is not a dictionary but a list of dictionaries.

[1]https://github.com/openstack/neutron/blob/86f94de99aa08b1b4aadca8e90c6e79487171b8e/neutron/policy.py#L142

Closes-Bug: #2098109
Change-Id: I592f7ff0673c15276e9da0054fd38f7ad96f795a