Remove ACCESS_READONLY RBAC constant.
This partially reverts commit [1] and removes constant ACCESS_READONLY
from the neutron_lib.constants module.
That constant was introduced as part of the RFE
https://bugs.launchpad.net/neutron/+bug/1875516 which was never fully
implemented and currently, with new S-RBAC API policies is not really
needed anymore.
It was decided during the Gazpacho PTG that this RFE will be closed and
already merged bits will be cleaned. That patch cleans it from the
neutron-lib repo.

[1] https://review.opendev.org/c/openstack/neutron-lib/+/730736

Depends-On: https://review.opendev.org/c/openstack/neutron/+/965928

Related-bug: #1875516

Change-Id: Iec8f1a3b3094c7d75882062f006d1efe77f71ca2
Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>