Service role is now considered as the one who can change project_id
Previously with [1] new property "can_set_project_id" was added to the
context.Context object but it was only True for the admin users
(is_admin=True) and if "_can_set_project_id" parameter was True
(configured via policy rules or passed directly to the __init__ method.

It had to be like that because "_is_service_role" wasn't defined properly
in the __init__ method and context.Context object wasn't able to be
created with "self.is_service_role" used in the "can_set_project_id"
property.
Due to that in the neutron_lib.api.attributes._validate_privileges
function it had to be explicitly checked if context is service role.

Now, as _is_service_role is defined properly in the
context.Context.__init__ method, it can be checked in the
"can_set_project_id" property and
neutron_lib.api.attributes._validate_privileges function can only check
if Context.can_set_project_id == True to allow changing project_id of
the resource.

[1] https://review.opendev.org/c/openstack/neutron-lib/+/969481

Related-bug: #2133212

Change-Id: Iacfaa22901af83bd72f418892771224f185c42e7
Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>